requirements.txt
matplotlib==3.8.4
Pillow==11.1.0
torch==2.3.0a0+6ddf5cf85e.nv24.4
torchvision==0.18.0a0
